agent/tools/web_scraper_mcp/requirements.txt
httpx==0.27.*
beautifulsoup4==4.12.*
fastmcp==2.13.*
python-dotenv==1.0.*
